Add `NoFinalize` marker trait
This trait gives the user a way to opt-out of finalization for certain
types when used with garbage collection. If a type, `T` implements
`NoFinalize`, this type will never be finalized when used inside `Gc` --
even if it has a destructor.

This trait is safe to implement because neither destructors nor
finalizers are guaranteed to be run in Rust, so deliberately preventing
finalization cannot, by definition, cause unsafety. That's not to say
that one should not be careful when implementing this trait, however:
doing so improperly can cause unintended memory leaks.