Add lint for traceable packed structs
Packed structs are not guaranteed to be aligned to a word boundary, so
they can't contain traceable fields (fields which could point to GC'd
values). This change adds a lint which runs as a late pass (after type
checking) in compilation.

The lint is set to error by default, but it can be disabled with
`#[allow(misaligned_gc_pointers)]`.