Add a way for the compiler to reference external `Gc` types
Currently `Gc` is defined in the libgc crate. The compiler isn't aware
of this so can't perform any specific operations on the `Gc` type. To
fix this, we introduce a `GcSmartPointer` trait which `libgc::Gc` will
implement.

This should be temporary. Eventually, the `Gc` smart pointer will be
part of the standard library. Right now though, it's convenient for
development that they are separated.