Add MIR pass to prevent early finalization
The optimizer (either in MIR or LLVM) may look at GC locals and
determine that their stack or register slot can be clobbered early if
there are no further uses of them. While such behaviour is legal and
conforms to the as-if rule, it can cause the collector finalizing values
earlier than expected, which has the potential to cause UAFs.

This adds a pass which extends the lifetime of GC locals in candidate
functions which are at risk of early finalization by the collector.
Candidate functions are those which contain local variables of type
`Gc<T>` which have a finalizer that will be run once it is unreachable.

This works by inserting the following asm to the end of the function,
(directly before the return statement) for each GC local whose lifetime
must not be shortened:

    llvm_asm!("" : : "r"(local_val) : "memory" : "volatile");

This pass is added early, right after borrow checking so that it happens
before other passes have a chance to shorten the lifetime of Gc locals.