Ensure `Collectable` trait method symbols are linked
After MIR lowering, rustc runs a "monomorphization collector" which
determines which monomorphized function symbols should be available
during the codegen phase. This can be thought of as a kind of tree
shaker which walks the CFG from a predetermined set of roots looking for
all instances of function calls. Those which are found are added to a
"mono-list".

The purpose of `std::mem::make_collectable<T>(value: &T)` is to recurse
through the fields of `value` and call `Collectable::set_collectable` on
all fields whose type implements `Collectable`. However, since the body
for this intrinsic is generated during codegen -- which is after the
mono collector has already run -- rustc won't know that calls to
`set_collectable` need to be kept in the binary.

This commit therefore patches the mono collector to do the following:

    * Look for all instances of `std::mem::make_collectable<T>`.

    * For each instance, recurse over the component types of `T` and add
      every component type's implementation of `Collectable` to the
      mono-list.