Allow make_collectable to be inlined
Intrinsics can't be lang_items so to get around this we hid it behind a
no-inline core library function of the same name so that it appeared in
the monomorphisation collector.

By switching to using a rustc_diagnostic_item, we don't need to use a
lang_item anymore and we can match on the intrinsic directly. This also
means that the first `make_collectable` call can be inlined away.