Remove `T: Send` requirement for `Gc<T>`
A `Gc<T>` no longer requires all values of `T` to impl `Send`. Instead,
this restriction is necessary only for certain methods, such as
`Gc::new`, which register an off-thread finalizer.

Methods such as `Gc::new` and `Gc::register_finalizer` are now defined
in impl blocks which require `T` to implement `Send`.

We also define a method, `Gc<T>::new_unfinalizable` which does not
require `T: Send`. This lets us use a `Gc` pointer for non sendable
types where the user signals explicitly that they are happy to omit a
finalizer. This is better for two reasons:

    1. The Gc API is more flexible and can be safely used with more
       types.
    2. It cleans up the function signatures for GC related methods. We
       no longer have to clutter types with `+ Send` everywhere.