Make `Gc<T>::new` require `T: Sync`
This fixes a soundness bug which can cause a race when the finalizer
thread and another thread can mutate a value without synchronisation.

This is because a `Gc` can point to another value which uses a different
shared ownership strategy (e.g. `Arc`) and then use a mutability
container like `RefCell` (which is `Send`) to update non-atomically. See
example below:

    struct Counter(Arc<RefCell<usize>>);

    impl Drop for Counter {
        fn drop(&mut self) {
            *self.0.borrow_mut() += 1;
        }
    }

    fn main() {
        let arc1 = Counter(Arc::new(RefCell::new(1)));
        {
            let gc1 = Gc::new(Arc::clone(arc1));
        }

        // Finalises 'gc1' off-thread
        force_gc();

        *arc1.borrow_mut() += 1;
        assert_eq!(arc1, 3);
    }