Add MIR pass to verify finalizers can be run safely
The GC runs finalizers on a separate thread, which means that values
with drop implementations must be Send + Sync. This puts a big
restriction on which types can be passed to Gc::new.

There are cases where Gc does not need to run finalizers for a value
because a) it doesn't have a drop method, or b) the drop call is removed
by finalizer optimisation. However, these values are still prevented
from being passed to Gc new because there's no way to check that T: Send
+ Sync + !Drop in the type system.

This pass allows such values to be constructed with Gc::new by removing
the T: Send + Sync bound on Gc::new, but then uses a MIR pass later in
the pipeline to check that if those values require dropping, they
implement Send + Sync.

We could allow more legal types to pass this check by moving it further
down into the codegen phase which happens after monomophisation, but
this would prevent it from running with a cargo check, so on balance I
think this is the right place. Values which violate this rule will lead
to a compiler error like below:

error: `Hello(123 as *mut u8)` cannot be safely finalized.
  --> src/main.rs:14:13
   |
14 |     Gc::new(Hello(123 as *mut u8));
   |     --------^^^^^^^^^^^^^^^^^^^^^-
   |     |       |
   |     |       has a drop method which cannot be safely finalized.
   |     `Gc::new` requires that it implements the `Send` trait.
   |
   = help: `Gc` runs finalizers on a separate thread, so `Hello(123 as *mut u8)` must implement `Send` in order to be safely dropped

error: `Hello(123 as *mut u8)` cannot be safely finalized.
  --> src/main.rs:14:13
   |
14 |     Gc::new(Hello(123 as *mut u8));
   |     --------^^^^^^^^^^^^^^^^^^^^^-
   |     |       |
   |     |       has a drop method which cannot be safely finalized.
   |     `Gc::new` requires that it implements the `Sync` trait.
   |
   = help: `Gc` runs finalizers on a separate thread, so `Hello(123 as *mut u8)` must implement `Sync` in order to be safely dropped