Add `FinalizerSafe` auto trait
A `Gc<T>` finalizes values on a separate thread. This means that if `T`
has a drop method, `T` must implement `Send` + `Sync` in order for drop
to be thread-safe.

This can be overly restrictive though, and users may unsafely impl
`Send` / `Sync` on a non-thread-safe `T`  to keep the compiler happy. At
first, it may seem fine to do this if they've verified the drop method
can never use `T` in an unsound way. However, making `T: Send + Sync`,
is a "global" hack, and it can now be erroneously used across threads
outside of a `Gc` - something the user did not intend.

This is where `FinalizerSafe` comes in, it can be thought of as making
`T: Send + Sync` but *only* when it's used in a finalizer. This is still
unsafe - and will need to be carefully checked each time - but it comes
with fewer unintended consequences.