Update `NonFinalizable` container
Sometimes it's necessary to prevent finalization on a per-use basis.
This can happen when it's not possible to implement `FinalizerOptional`
on types because were defined outside of the current crate (due to the
orphan rule).

This updates the `NonFinalizable` container to respect the new semantics
and adds some tests for it.