Fix soundness hole in GC finalisers
A potential soundness hole exists in GC finalisers where `Gc<T>`s that
contain ordinary Rust references (&/&mut).

Consider the following:

```rust
fn main() {
    let b = Box::new(123);
    let gc: Gc<&usize> = Gc::new(b.as_ref());

    // foo(b);
    // dbg(gc); // ERROR: use-after-move
}
```

This kind of program might seem weird, but it is perfectly safe. The lifetime of
`gc` is tied to `b`. The compiler catches uses of `gc` after a move of `b` as
we'd expect.

If we introduce finalisers into the mix, this can become unsound. It seems like
ownership should save us here, after all, `gc` cannot outlive `b`. However,
finalisers provide a stealthy way to violate soundness because they can run at
the collector's leisure - potentially long after `b` has been dropped. Let's
extend the above example to make this more concrete:

```rust

struct Wrapper<'a>(&'a usize);

impl<'a> Drop for Wrapper<'a> {
    fn drop(&mut self) {
        println!("Dropping {}", self.0);
    }
}

fn main() {

    {
        let b = Box::new(123);
        let gc = Gc::new(Wrapper(b.as_ref()));
    } // b is dropped

    // GC happens here, calling `Wrapper::drop` as a finaliser.
    // BOOM! use-after-free.
}
```

The above can cause a use-after-free if the finaliser for `Wrapper` (which
derefs the `&usize` field) is run after `b`'s `Box` is already droppped.

Fortunately there is a solution to this. It's possible by imposing a rule on
`Gc<T>`s:

    A type `T` must never be placed in a `Gc<T>` if the both of the following
    statements are true for `T`:

        1. `T` (or any component type of `T`) contains any field of type `&U` or
        `&mut U`.

        2. `T` requires a finaliser.

This rule is fairly easy to enforce with a slight modification to our existing
finaliser checks. We first introduce a new auto trait `ReferenceFree`, which is
unimplemented on reference types:

```rust
impl !ReferenceFree for &T {}
impl !ReferenceFree for &mut T {}
```

Then, at every `Gc<T>::new` callsite, the compiler will check whether `T`
adheres to the new rule. If it doesn't, we throw a compiler error. The following
shows the compiler error for the second example above:

```
error: `Wrapper(b.as_ref())` cannot be safely constructed.
  --> src/main.rs:19:26
   |
19 |         let gc = Gc::new(Wrapper(b.as_ref()));
   |                  --------^^^^^^^^^^^^^^^^^^^-
   |                  |       |
   |                  |       contains a reference (&) which may no longer be valid when it is finalized.
   |                  `Gc::new` requires that a type is reference free.

warning: `ref_soundness` (bin "ref_soundness") generated 1 warning
error: could not compile `ref_soundness` due to previous error; 1 warning emitted
```

The drawback of this approach is that it can be overly restrictive. This will
disallow `T`s with reference fields which are never actually used inside
`T::drop`. I believe a more advanced version of drop MIR checking could account
for this, but it's tricky. We would need a way to differentiate between
perfectly safe references like `&mut self`, and unsafe references to other
objects. This is not easy in MIR with its current limited form of def-use
analysis.

Instead, I feel the current approach is adequate. If a user _really_ wants to
use `T` with reference fields in a `Gc` and they are certain that it is not
being used in `T::drop` in an unsound way, then they can always unsafe impl
`ReferenceFree` for `T`, and `Gc<T>` would pass the analysis.