Change FSA to check for `Send + Sync` explicitly
Previously FSA only checked that drop methods dereferenced `FinalizerSafe`
fields. The benefit of this is that it allows a greater number of
finalisers to pass because `FinalizerSafe` is concerned with type safety
in the context of GC only.

However, as an auto trait, `FinalizerSafe` is automatically implemented
on types in third party crates which have no knowledge of Alloy. This
could mean that non-Send + non-Sync types are erroneously marked as safe
to finalise. In the interest of soundness, we err on the side of caution
and modify FSA to check for the existence of `FinalizerSafe + Send +
Sync`.