Do not suppress all Boehm warnings
If something goes wrong during a GC, we likely want to know what
happened.

It's not totally clear whether this is the correct way to disable
warnings anyway (it uses some deprecated library features in Boehm).
Annoying warnings (such as those for many large page allocations) can be
disabled individually.