Use GC_memalign for larger allocations
When a layout's alignment is larger than its size, we must allocate
using Boehm's `memalign` function instead of `malloc` in order to mirror
the existing behaviour in rustc.

This is checked in rustc with various assertions on pointer read and
write intrinsics, and was discovered when they failed with large
alloctions in lrlex using a debug build of alloy.