Disable FSA for the std::gc crate
In some of its methods, the `gc` standard library crate creates
temporary GC objects. However, when FSA is run on this it doesn't have
access to the monomorphised types so is too overly cautious.

Since we control the GC crate, and we are deep in unsafe territory here
anyway, we should disable FSA for the GC crate.